+++ .changeset/smooth-cameras-prove.md
@@ -1,0 +1,5 @@
+---
+"@primer/react": major
+---
+
+Move deprecated components to deprecated folder
