+++ deprecated/package.json
@@ -1,0 +1,9 @@
+{
+    "_comment1": "this is required only for typescript. once this is fixed https://github.com/microsoft/TypeScript/issues/33079 we can remove this hack",
+    "name": "@primer/react/deprecated",
+    "types": "../lib-esm/deprecated/index.d.ts",
+    "main": "../lib-esm/deprecated/index.js",
+    "type": "module",
+    "sideEffects": false
+  }
+
