+++ package.json
@@ -7,11 +7,18 @@
   "exports": {
     ".": {
       "node": "./lib/index.js",
+      "require": "./lib/index.js",
       "default": "./lib-esm/index.js"
     },
     "./drafts": {
       "node": "./lib/drafts/index.js",
+      "require": "./lib/drafts/index.js",
       "default": "./lib-esm/drafts/index.js"
+    },
+    "./deprecated": {
+      "node": "./lib/deprecated/index.js",
+      "require": "./lib/deprecated/index.js",
+      "default": "./lib-esm/deprecated/index.js"
     },
     "./lib-esm/*": {
       "node": [
