+++ src/__tests__/BorderBox.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import theme from '../theme'
-import {BorderBox} from '..'
+import {BorderBox} from '../deprecated'
 import {render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
