+++ src/__tests__/Dropdown.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Dropdown} from '..'
+import {Dropdown} from '../deprecated'
 import {behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
