+++ src/__tests__/Grid.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Grid} from '..'
+import {Grid} from '../deprecated'
 import {render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
