+++ src/__tests__/Position.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import {Box, Position, Absolute, Fixed, Relative, Sticky} from '..'
+import {Box} from '..'
+import {Position, Absolute, Fixed, Relative, Sticky} from '../deprecated'
 import {render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
